+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
  * added `NullDiscovery`
  * changed `ResourceDiscovery::find()` and `getBindings()` to throw a
    `NoSuchTypeException` if the type has not been defined
+ * removed dependency to beberlei/assert
 
 * 1.0.0-beta (2015-01-12)